refactor nav in to default view
