+++ app/styles/nav.scss
@@ -1,5 +1,5 @@
 $mobile-nav-height: 60px;
-$other-nav-height: 60px;
+$other-nav-height: 80px;
 
 .nav-bar {
 
@@ -10,7 +10,7 @@ $other-nav-height: 60px;
   height: $other-nav-height;
 
   .page-name {
-    margin-top: $other-nav-height / 5;
+    margin-top: $other-nav-height / 3;
     text-align: center;
     font-size: 36px;
   }
@@ -33,6 +33,7 @@ $other-nav-height: 60px;
 
 }
 
+
 .nav-burger {
 
   $height: 40px;
@@ -44,8 +45,14 @@ $other-nav-height: 60px;
   width: $width;
   height: $height;
   background-color: $rad-red;
-  margin-left: 10px;
-  margin-top: 10px;
+
+  margin-left: 20px;
+  margin-top: 20px;
+
+  @media #{$sm} {
+    margin-left: 10px;
+    margin-top: 10px;
+  }
 
   &:hover {
     background-color: $desat-rad-red;
