+++ app/templates/default-layout.hbs
@@ -1,5 +1,4 @@
 <nav class="nav-bar">
-  
   <div class="container">
     <div class="grid">
       <div class="row">
@@ -10,7 +9,7 @@
             <div class="nav-bun"></div>
           </div>
         </div>
-        <h2 class="page-name g-6-12">Leader</h2>
+        <h2 class="page-name g-6-12">{{view.pageTitle}}</h2>
       </div>
     </div>
   </div>
