+++ tests/unit/views/default-layout-test.js
@@ -1,0 +1,12 @@
+import {
+  moduleFor,
+  test
+} from 'ember-qunit';
+
+moduleFor('view:default-layout');
+
+// Replace this with your real tests.
+test('it exists', function(assert) {
+  var view = this.subject();
+  assert.ok(view);
+});
